Make sure that an ORM sync is executed when it is requested

+++ src/gateway/base_controller.py
@@ -48,17 +48,17 @@ class BaseController(object):
         self._maintenance_controller = maintenance_controller  # type: MaintenanceController
         self._sync_orm_thread = None  # type: Optional[DaemonThread]
         self._master_controller.subscribe_event(self._handle_master_event)
-        self._maintenance_controller.subscribe_maintenance_stopped(self.sync_orm)
+        self._maintenance_controller.subscribe_maintenance_stopped(self.request_sync_orm)
         self._sync_orm_interval = sync_interval
+        self._sync_running = False
 
     def _handle_master_event(self, master_event):  # type: (MasterEvent) -> None
         if master_event.type in [MasterEvent.Types.EEPROM_CHANGE, MasterEvent.Types.MODULE_DISCOVERY]:
-            if self._sync_orm_thread is not None:
-                self._sync_orm_thread.request_single_run()
+            self.request_sync_orm()
 
     def start(self):
         self._sync_orm_thread = DaemonThread(name='ORM syncer for {0}'.format(self.__class__.__name__),
-                                             target=self.sync_orm,
+                                             target=self._sync_orm,
                                              interval=self._sync_orm_interval,
                                              delay=300)
         self._sync_orm_thread.start()
@@ -67,27 +67,46 @@ class BaseController(object):
         if self._sync_orm_thread is not None:
             self._sync_orm_thread.stop()
 
-    def sync_orm(self):
+    def request_sync_orm(self):
+        if self._sync_orm_thread is not None:
+            self._sync_orm_thread.request_single_run()
+
+    def run_sync_orm(self):
+        self._sync_orm()
+
+    def _sync_orm(self):
+        # type: () -> bool
         if self.SYNC_STRUCTURES is None:
-            return
+            return False
 
-        for structure in self.SYNC_STRUCTURES:
-            orm_model = structure.orm_model
-            try:
-                name = structure.name
-                skip = structure.skip
+        if self._sync_running:
+            for structure in self.SYNC_STRUCTURES:
+                orm_model = structure.orm_model
+                logger.info('ORM sync ({0}): Already running'.format(orm_model.__name__))
+            return False
+        self._sync_running = True
 
-                logger.info('ORM sync ({0})'.format(orm_model.__name__))
+        try:
+            for structure in self.SYNC_STRUCTURES:
+                orm_model = structure.orm_model
+                try:
+                    name = structure.name
+                    skip = structure.skip
 
-                ids = []
-                for dto in getattr(self._master_controller, 'load_{0}s'.format(name))():
-                    if skip is not None and skip(dto):
-                        continue
-                    id_ = dto.id
-                    ids.append(id_)
-                    orm_model.get_or_create(number=id_)  # type: ignore
-                orm_model.delete().where(orm_model.number.not_in(ids)).execute()  # type: ignore
+                    logger.info('ORM sync ({0})'.format(orm_model.__name__))
 
-                logger.info('ORM sync ({0}): completed'.format(orm_model.__name__))
-            except Exception:
-                logger.exception('ORM sync ({0}): Failed'.format(orm_model.__name__))
+                    ids = []
+                    for dto in getattr(self._master_controller, 'load_{0}s'.format(name))():
+                        if skip is not None and skip(dto):
+                            continue
+                        id_ = dto.id
+                        ids.append(id_)
+                        orm_model.get_or_create(number=id_)  # type: ignore
+                    orm_model.delete().where(orm_model.number.not_in(ids)).execute()  # type: ignore
+
+                    logger.info('ORM sync ({0}): completed'.format(orm_model.__name__))
+                except Exception:
+                    logger.exception('ORM sync ({0}): Failed'.format(orm_model.__name__))
+        finally:
+            self._sync_running = False
+        return True
